// src/app/model/asset.ts
export class asset {
    symbol: string;
    shares:        any = 0;
    avgprice:      any = 0;
    sharesSold:    any= 0;
    avgpriceSold:  any= 0;
    originalMoney: any= 0;
    totalMoneyIn:  any= 0;
    totalMoneyOut: any= 0;
    price:         any= 0;
    currentTotal:  any= 0;
    realProfit:    any= 0;
    realMargin:    any= 0;
    unRealProfit:  any= 0;
    unRealMargin:  any= 0;
}